requirements.txt
 streamlit 
 youtube-transcript-api 
 cohere 
 python-dotenv

